Private types and functions

Documentation for ClimaSeaIce.jl's internal interfaces.

ClimaSeaIce

ClimaSeaIce.LinearLiquidus — Type

LinearLiquidus(FT=Float64,
               slope = 0.054, # psu / ᵒC
               freshwater_melting_temperature = 0) # ᵒC

Return a linear model for the dependence of the melting temperature of saltwater on salinity,

\[Tₘ(S) = T₀ - m S ,\]

where $Tₘ(S)$ is the melting temperature as a function of salinity $S$, $T₀$ is the melting temperature of freshwater, and $m$ is the ratio between the melting temperature and salinity (in other words the linear model should be thought of as defining $m$ and could be written $m ≡ (T₀ - Tₘ) / S$. The signs are arranged so that $m > 0$ for saltwater).

The defaults assume that salinity is given in practical salinity units psu and temperature is in degrees Celsius.

Note: the function melting_temperature(liquidus, salinity) returns the melting temperature given salinity.

ClimaSeaIce.PhaseTransitions — Type

PhaseTransitions(FT=Float64,
                 ice_density           = 917,   # kg m⁻³
                 ice_heat_capacity     = 2000,  # J / (kg ᵒC)
                 liquid_density        = 999.8, # kg m⁻³
                 liquid_heat_capacity  = 4186,  # J / (kg ᵒC)
                 reference_latent_heat = 334e3  # J kg⁻³
                 liquidus = LinearLiquidus(FT)) # default assumes psu, ᵒC

Return a representation of transitions between the solid and liquid phases of salty water: in other words, the freezing and melting of sea ice.

The latent heat of fusion $ℒ(T)$ (more simply just "latent heat") is a function of temperature $T$ via

\[ρᵢ ℒ(T) = ρᵢ ℒ₀ + (ρ_ℓ c_ℓ - ρᵢ cᵢ) (T - T₀) \]

where $ρᵢ$ is the ice_density, $ρ_ℓ$ is the liquid density, $cᵢ$ is the heat capacity of ice, and $c_ℓ$ is the heat capacity of liquid, and $T₀$ is a reference temperature, all of which are assumed constant.

The default liquidus assumes that salinity has practical salinity units (psu) and that temperature is degrees Celsius.

ClimaSeaIce.EnthalpyMethodSeaIceModels

ClimaSeaIce.EnthalpyMethodSeaIceModels.EnthalpyMethodSeaIceModel — Method

EnthalpyMethodSeaIceModel(; grid, kw...)

Return a thermodynamic model for ice sandwiched between an atmosphere and ocean on an Eulerian grid.

ClimaSeaIce.EnthalpyMethodSeaIceModels.compute_tendencies! — Method

Calculate the right-hand-side of the free surface displacement (η) equation.